recipe-scraper/pyproject.toml
[tool.poetry]
name = "recipe-scraper"
version = "0.1.0"
description = ""
authors = ["Lu Wang <luwang.wang@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
recipe-scrapers = "^15.0.0"
fastapi = {extras = ["standard"], version = "^0.112.0"}
openai = "^1.40.6"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
